+++ Assets/Palette.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 53eac4ead7d9b0d40a8f39e6da12a5c4
-PrefabImporter:
+guid: 52202566d936b6f4f8e4baa58c7dbcb5
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
