+++ Assets/PhysicsMaterials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7ac1aeb24e6c8d1488d6174f344ab1c7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
