+++ Assets/Prefabs/SpawnPoint.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fbfc5d25bda5c8f4a9a97786f13a0da0
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
