+++ Assets/Prefabs/Star.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 70c9bc53328b7c54cae564f6bfb25ce7
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
