+++ Assets/Resources.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 9fcf1c11052bb1f4c89d90c1666c8a95
-NativeFormatImporter:
+guid: 4d71d8e0376f3ef46b688c76830ac387
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
