+++ ProjectSettings/TagManager.asset
@@ -17,7 +17,7 @@ TagManager:
   - UI
   - WidthBooster
   - HeightBooster
-  - 
+  - UPBooster
   - 
   - 
   -
